Moved create_plots into plot.py
