+++ pysrc/utils/plots.py
@@ -1,8 +1,24 @@
+import os
+
 import seaborn as sns
 import matplotlib
+import matplotlib.pyplot as plt
 import pandas as pd
+from typing import Callable
 
-
+def create_plot(dataframe: pd.DataFrame, groupby_attrs: list[str], plot_method: Callable, plot_folder: str, plot_fname: str, figsize = (15,10)):
+    f, axes = plt.subplots(1, len(groupby_attrs), figsize=figsize)
+    sns.despine(f)
+    
+    if len(groupby_attrs) > 1:
+        for ax, groupby_attr in zip(axes, groupby_attrs):
+            ax = plot_method(axes = ax, dataframe = dataframe, groupby_attr = groupby_attr)
+    else:
+        axes = plot_method(axes = axes, dataframe = dataframe, groupby_attr = groupby_attrs[0])
+    
+    plot_path = os.path.join(os.getcwd(), plot_folder, plot_fname)
+    f.savefig(plot_path)
+    return plot_path
 
 def create_PrecisionRecallF1_lineplot(axes : matplotlib.axes.Axes, dataframe: pd.DataFrame, groupby_attr: str) -> matplotlib.axes.Axes:
     df_grouped = dataframe.groupby(groupby_attr)
